MAJOR: Pagination Support
This commit is what I was afraid of doing during the alpha release, delivering a bunch of API Calls and then having to refactor them all...

 MAJOR breaking change from the Alpha Release as unstructured data is no longer passed through functions using generic interface{} (yay!), and the API/session/session function abstracted by a new Sessions function, which gives us Pagination support.

Significantly noteworthy is the enhancements to Verbosity that came about as a side effect of the API/Result data structure. When called using the -v flag, you have access to full HTTP Request & Response metadata, which helps a lot when troubleshooting API calls.

More to come soon!