feat: usable pattern for flags
Found a pattern that works. Right now all the flags are global, would be nice to sort the output of the cli help documentation from the actual flags that should be global and those which are subcommand specific but have yet to find a solution. For now unblocked on provisioning get commands.